requirements.txt
streamlit
tensorflow==2.13.0
opencv-python-headless
numpy
Pillow
nltk
scikit-learn